reorganiza css e corrige erro nas tooltips
