+++ public/css/form-acesso.css
@@ -6,10 +6,11 @@
 
 a {
   color: hsl(213, 52%, 52%);
+  
 }
 
 a:hover {
-  color: var(--secondary-gray);
+  color: rgb(10, 88, 202);
 }
 
 /* LEFT PANEL */
@@ -179,6 +180,7 @@ a:hover {
   --bs-tooltip-opacity: 1;
   --bs-tooltip-padding-x: .8rem;
   --bs-tooltip-padding-y: .4rem;
+  --bs-tooltip-max-width: 300px;
   margin-bottom: 3px !important;
 }
 
@@ -188,7 +190,7 @@ a:hover {
 
 .form-control.is-invalid:focus {
   box-shadow: none;
-  outline: 1px solid rgba(233, 201, 201, 0.712);
+  outline: 1px solid rgba(233, 201, 201, 0.7);
 }
 
 /* MODAL */
@@ -221,7 +223,7 @@ a:hover {
   .photo-logo {
     width: 100px;
     height: 100px;
-    margin: 0 -20px;
+    margin: 0 -18px;
   }
 }
 
@@ -246,6 +248,10 @@ a:hover {
   .photo-logo {
     width: 80px;
     height: 80px;
-    margin: 10px -16px;
+    margin: 10px -15px;
+  }
+
+  .custom-tooltip {
+    --bs-tooltip-max-width: 80vw;
   }
 }
