+++ public/html/cadastro-cliente.html
@@ -12,7 +12,7 @@
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css">
 
   <!-- CSS Projeto -->
-  <link rel="stylesheet" href="../css/cadastro-global.css">
+  <link rel="stylesheet" href="../css/form-acesso.css">
   <link rel="shortcut icon" href="../assets/images/favicon.png" type="image/x-icon">
 </head>
 
@@ -24,7 +24,7 @@
         <div class="panel-initial text-white d-flex flex-column justify-content-center h-100 w-100 gap-2 gap-sm-3">
           <img class="photo-logo" src="../assets/images/logo-white.png" alt="Logo">
           <div>
-            <h2 class="fw-bold mb-3">Bem-vindo à Habitta!</h2>
+            <h2 class="fw-bold mb-3">Bem-vindo à Habitta</h2>
             <p class="mb-4 description-text">
               Aqui você encontra os melhores profissionais para os seus serviços, de forma rápida e segura.
             </p>
@@ -48,17 +48,6 @@
               <input class="btn-check" type="radio" name="type" id="pessoaJuridica" value="B2B">
               <label class="btn btn-outline-custom" for="pessoaJuridica">Pessoa Jurídica</label>
             </div>
-            <!-- <div class="mb-3">
-              <div class="form-check form-check-inline">
-                <input class="form-check-input" type="radio" name="type" id="pessoaFisica" value="B2C" checked>
-                <label class="form-check-label" for="pessoaFisica">Pessoa Física</label>
-              </div>
-              <div class="form-check form-check-inline">
-                <input class="form-check-input" type="radio" name="type" id="pessoaJuridica" value="B2B">
-                <label class="form-check-label" for="pessoaJuridica">Pessoa Jurídica</label>
-              </div>
-            </div> -->
-
             <!-- Nome / Razão Social -->
             <div class="mb-3">
               <input type="text" class="form-control input-custom" id="nameInput" name="name" placeholder="Nome completo" required>
@@ -95,10 +84,10 @@
           </div>
           <!-- Suporte -->
           <div class="social-login text-center mt-4">
-            <a href="https://wa.me/5553991962807" class="btn btn-social me-2">
+            <a href="https://wa.me/5553991962807" class="btn btn-social me-2" title="Suporte via WhatsApp">
               <i class="bi bi-whatsapp"></i>
             </a>
-            <a href="mailto:habittaoficial@gmail.com" target="_blank" rel="noopener noreferrer" class="btn btn-social">
+            <a href="mailto:habittaoficial@gmail.com" target="_blank" class="btn btn-social" title="Suporte via e-mail">
               <i class="bi bi-envelope"></i>
             </a>
           </div>
