+++ public/html/cadastro.html
@@ -31,12 +31,15 @@
   </header>
 
   <main class="auth-container">
-    <div class="container">
-      <h2 class="fw-bold title text-white">Como você gostaria de se cadastrar?</h2>
-      <p class="subtitle">Escolha a opção que melhor descreve o seu perfil.</p>
+    <div class="container my-5">
       <div class="row g-4 justify-content-center">
+        <div class="col-12 mt-4 m-lg-0">
+          <h2 class="fw-bold title text-white">Como você gostaria de se cadastrar?</h2>
+          <p class="subtitle">Escolha a opção que melhor descreve o seu perfil.</p>
+        </div>
+
         <!-- Cliente -->
-        <div class="col-12 col-md-6 col-lg-4">
+        <div class="col-12 col-md-6 col-lg-5 col-xl-4">
           <div class="card card-option h-100">
             <div class="icon">
               <i class="bi bi-person"></i>
@@ -56,7 +59,7 @@
         </div>
 
         <!-- Prestador -->
-        <div class="col-12 col-md-6 col-lg-4">
+        <div class="col-12 col-md-6 col-lg-5 col-xl-4">
           <div class="card card-option h-100">
             <div class="icon">
               <i class="bi bi-briefcase"></i>
